bug(acumen_product_agent, acumen_product_query_concern): fix category
fix categories as well as change sync to only run master products that
are active through the sync, other products will still be synced but
they will be variants on master products. If there is any master
products that are not active products and not variants on an active
master product they will not be synced anymore.