build: fix the build by using an old version of huginn and adding a docker-compose
The current version of huginn_agent, v0.6.1 released in 2017, isn't compatible with the most recent version of huginn.
Using the 5 Stones fork, this locks the version to the one that the helm chart deploys by default, and updates rake and ruby to match.
The rake command also requires a DB, which is covered using a docker-compose instead of a docker run command.